admin/requirements.txt
Django==5.2.7
psycopg2-binary==2.9.11
python-decouple==3.8
django-extensions==4.1
django-crispy-forms==2.3
crispy-bootstrap5==0.7
django-tables2==2.7.0
django-filter==24.3
django-import-export==4.1.1
